removed tests on time execution (hard to reproduce in different environments live CI systems)
